Move import_stage which calls get_package_dict into DguHarvesterBase, so that we are more likely to merge with ckan/ckanext-harvest.
