+++ setup.py
@@ -33,8 +33,8 @@ setup(
     [ckan.plugins]
 	# Add plugins here, eg
 	harvest=ckanext.harvest.plugin:Harvest
-	ckan_harvester=ckanext.harvest.harvesters:CKANHarvester
-	dkan_harvester=ckanext.harvest.harvesters:DKANHarvester
+	ckan_harvester=ckanext.harvest.harvesters.ckanharvester:CKANHarvester
+    dkan_harvester=ckanext.harvest.harvesters.dkanharvester:DKANHarvester
 	[paste.paster_command]
 	harvester = ckanext.harvest.commands.harvester:Harvester
 	""",
